Update greeter to move ApplicationComponent into shared modules
- There's one version for android in shared/androidMain
- There's another version of iOS in shared/iosMain
- expect/actual is used to reference the generated code in multiple iOS
targets
