+++ multiplatform/greeter/androidApp/src/main/java/com/fredporciuncula/inject/greeter/android/MainActivity.kt
@@ -11,6 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import com.fredporciuncula.inject.greeter.ApplicationComponent
 import com.fredporciuncula.inject.greeter.CommonGreeter
 import me.tatarka.inject.annotations.Component
