+++ multiplatform/greeter/shared/build.gradle.kts
@@ -54,6 +54,7 @@ dependencies {
   // KSP will eventually have better multiplatform support and we'll be able to simply have
   // `ksp libs.kotlinInject.compiler` in the dependencies block of each source set
   // https://github.com/google/ksp/pull/1021
+  add("kspAndroid", libs.kotlinInject.compiler)
   add("kspIosX64", libs.kotlinInject.compiler)
   add("kspIosArm64", libs.kotlinInject.compiler)
   add("kspIosSimulatorArm64", libs.kotlinInject.compiler)
