+++ multiplatform/greeter/shared/src/iosArm64Main/kotlin/com/fredporciuncula/inject/greeter/ApplicationComponent.iosArm64.kt
@@ -1,0 +1,4 @@
+package com.fredporciuncula.inject.greeter
+
+actual fun ApplicationComponent.Companion.create(): ApplicationComponent =
+  ApplicationComponent::class.create()
