--- multiplatform/greeter/shared/src/iosSimulatorArm64Main/kotlin/com/fedporciuncula/inject/greeter/ApplicationComponent.kt
@@ -1,12 +1,0 @@
-package com.fredporciuncula.inject.greeter
-
-import me.tatarka.inject.annotations.Component
-
-@Component
-abstract class ApplicationComponent : PlatformComponent {
-  abstract val greeter: CommonGreeter
-
-  companion object {
-    fun create() = ApplicationComponent::class.create()
-  }
-}
