+++ multiplatform/greeter/shared/src/iosSimulatorArm64Main/kotlin/com/fredporciuncula/inject/greeter/ApplicationComponent.iosSimulatorArm64.kt
@@ -1,0 +1,4 @@
+package com.fredporciuncula.inject.greeter
+
+actual fun ApplicationComponent.Companion.create(): ApplicationComponent =
+  ApplicationComponent::class.create()
