+++ multiplatform/greeter/shared/src/iosX64Main/kotlin/com/fredporciuncula/inject/greeter/ApplicationComponent.iosX64.kt
@@ -1,0 +1,4 @@
+package com.fredporciuncula.inject.greeter
+
+actual fun ApplicationComponent.Companion.create(): ApplicationComponent =
+  ApplicationComponent::class.create()
